refactor(cli): use jsonrpc schemas
